Add swiftpm 4 package manifest
As recommended in the [Version-specific manifest selection][manifest-version]
section of the swiftpm docs, the `Package.swift` file (without any specific
version suffix) should target the latest stable version of Swift.

This renames the existing `Package.swift` file to `Package@swift-3.swift`,
and adds a new and updated version for Swift 4.

[manifest-version]: https://github.com/apple/swift-package-manager/blob/31e5affada8e354b66d3bda64f6e4245f5da5446/Documentation/Usage.md#version-specific-manifest-selection
