+++ .swift-version
@@ -1,1 +1,1 @@
-3.1
+4.0
